% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_concordance.R
\name{concordance}
\alias{concordance}
\title{Concordance analysis}
\usage{
concordance(file1, file2, haplotypes = FALSE)
}
\arguments{
\item{file1}{is either an xlsx or csv file}

\item{file2}{is either an xlsx or csv file}

\item{haplotypes}{(required) if set to TRUE, will no longer change the allele arrangement.}
}
\description{
This package identifies overlapping data (usually samples) between two files 
and calculates the total number of matching data (usually genotypes). 
It creates both an output csv file and a barplot.
}
\examples{
concordance("genotypes1.xlsx", "genotypes2.xlsx", haplotypes = FALSE)
}
